+++ app/src/test/java/org/simple/clinic/summary/PatientSummaryInitTest.kt
@@ -15,9 +15,7 @@ class PatientSummaryInitTest {
   private val patientUuid = UUID.fromString("fca8c3ad-75ca-4053-ba2f-e5c8ffda8991")
   private val defaultModel = PatientSummaryModel.from(ViewExistingPatient, patientUuid)
   private val facility = TestData.facility(uuid = UUID.fromString("6af0d0c5-4c5d-493f-bba9-0bcffde27e07"))
-  private val initSpec = InitSpec(PatientSummaryInit(
-      isCdsAlertsFeatureEnabled = true
-  ))
+  private val initSpec = InitSpec(PatientSummaryInit())
 
   @Test
   fun `when the screen is created, load the initial data`() {
@@ -32,25 +30,7 @@ class PatientSummaryInitTest {
                     CheckForInvalidPhone(patientUuid),
                     LoadMedicalOfficers,
                     LoadPatientRegistrationData(patientUuid),
-                    LoadClinicalDecisionSupportInfo(patientUuid)
-                )
-            )
-        )
-  }
-
-  @Test
-  fun `when the screen is created and cds alerts feature is disabled, then do not load cds info`() {
-    initSpec
-        .whenInit(defaultModel)
-        .then(
-            assertThatFirst(
-                hasModel(defaultModel),
-                hasEffects(
-                    LoadPatientSummaryProfile(patientUuid),
-                    LoadCurrentUserAndFacility,
-                    CheckForInvalidPhone(patientUuid),
-                    LoadMedicalOfficers,
-                    LoadPatientRegistrationData(patientUuid)
+                    CheckIfCDSSPilotIsEnabled
                 )
             )
         )
@@ -83,7 +63,7 @@ class PatientSummaryInitTest {
                 hasModel(model),
                 hasEffects(
                     LoadPatientSummaryProfile(patientUuid) as PatientSummaryEffect,
-                    LoadClinicalDecisionSupportInfo(patientUuid)
+                    CheckIfCDSSPilotIsEnabled
                 )
             )
         )